Remove superfluous functions
I'm pretty sure the new data structure makes these functions trivial. I
had these functions in so that I could scale the x-limits of all the
histogram plots identically. In order to do this scaling, I had to find
the menu item of the brand that was largest. Now that everything is in a
flat data structure, I can simply extract the values corresponding to
the key I want (e.g. nf_sugars), and determine the max from the
resulting list.